+++ evolution-go-manager/src/services/api/instances.ts
@@ -372,6 +372,7 @@ export interface ChatwootConfig {
   importMessages: boolean;
   daysLimitImportMessages: number;
   autoCreate: boolean;
+  ignoreGroups?: boolean;
   inboxId?: string;
 }